get tyr building after basically gutting it
