+++ Makefile.am
@@ -37,28 +37,19 @@ endif
 
 # lib valhalla compilation etc
 lib_LTLIBRARIES = libvalhalla_tyr.la
-nobase_include_HEADERS = \
-	valhalla/tyr/handler.h \
-	valhalla/tyr/route_handler.h \
-	valhalla/tyr/custom_route_handler.h \
-	valhalla/tyr/locate_handler.h \
-	valhalla/tyr/nearest_handler.h
-
-libvalhalla_tyr_la_SOURCES = \
-	src/tyr/handler.cc \
-	src/tyr/route_handler.cc \
-	src/tyr/custom_route_handler.cc \
-	src/tyr/locate_handler.cc \
-	src/tyr/nearest_handler.cc
+nobase_include_HEADERS = valhalla/tyr/service.h
+libvalhalla_tyr_la_SOURCES = src/tyr/service.cc
 libvalhalla_tyr_la_CPPFLAGS = $(DEPS_CFLAGS) $(VALHALLA_CPPFLAGS) @BOOST_CPPFLAGS@
 libvalhalla_tyr_la_LIBADD = $(DEPS_LIBS) $(VALHALLA_LDFLAGS) @BOOST_LDFLAGS@ $(BOOST_SYSTEM_LIB) $(BOOST_THREAD_LIB)
 
 #distributed executables
-bin_PROGRAMS = tyr_simple_service
-tyr_simple_service_SOURCES = \
-        src/tyr/simple_service.cc
+bin_PROGRAMS = tyr_simple_service tyr_service
+tyr_simple_service_SOURCES = src/tyr/simple_service.cc
 tyr_simple_service_CPPFLAGS = $(DEPS_CFLAGS) $(VALHALLA_CPPFLAGS) @BOOST_CPPFLAGS@
 tyr_simple_service_LDADD = $(DEPS_LIBS) $(VALHALLA_LDFLAGS) @BOOST_LDFLAGS@ $(BOOST_SYSTEM_LIB) $(BOOST_THREAD_LIB) libvalhalla_tyr.la
+tyr_service_SOURCES = src/tyr/tyr_service.cc
+tyr_service_CPPFLAGS = $(DEPS_CFLAGS) $(VALHALLA_CPPFLAGS) @BOOST_CPPFLAGS@
+tyr_service_LDADD = $(DEPS_LIBS) $(VALHALLA_LDFLAGS) @BOOST_LDFLAGS@ $(BOOST_SYSTEM_LIB) $(BOOST_THREAD_LIB) libvalhalla_tyr.la
 
 # tests
 check_PROGRAMS = \
